Lab3 initial commit

+++ src/main/java/AirlineManagement/Flight.java
@@ -43,7 +43,7 @@ public class Flight {
         return id;
     }
     
-    public float getPrice()
+    public double getPrice()
     {
         return this.price;
     }
